// src/api/generated/models/paginatedDistrictsResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * ECD Backend API
 * Early Childhood Development management system API.
 * OpenAPI spec version: 1.0
 */
import type { DistrictResponseDto } from './districtResponseDto';

export interface PaginatedDistrictsResponseDto {
  items: DistrictResponseDto[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
}